re-organize build common util
