+++ tvm/src/codegen/build_common.h
@@ -12,6 +12,7 @@
 #include "../runtime/meta_data.h"
 
 namespace TVM {
+
 namespace codegen {
 // Extract function information from device function.
 inline std::unordered_map<std::string, runtime::FunctionInfo>
